refactor(user-role):seperate validation from middleware
