+++ src/__tests__/unknownRoute.js
@@ -21,7 +21,7 @@ describe('EMAIL ROUTE', () => {
       const response = await request.post('/');
 
       expect(response.status).to.equal(404);
-      expect(response.body.errors.message).to.equal('Not Found');
+      expect(response.body.message).to.equal('Not Found');
     });
   });
 });
