+++ src/index.js
@@ -34,7 +34,11 @@ if (!isProduction) {
 app.use(routes);
 
 // / catch 404 and forward to error handler
-
+app.use((req, res, next) => {
+  const err = new Error('Not Found');
+  err.status = 404;
+  next(err);
+});
 
 // / error handlers
